Linearize derivatives at the same timepoints as primals
After much discussion and investigation, it is decided that there is not
an easy way to determine a good `abstol`/`reltol` to use for separate
linearization of derivatives, and as such there is no reasonable rule of
thumb for knowing how many timepoints is "enough" for the derivatives.

We instead make the same choice as the base solvers themselves; we
sample the derivatives at precisely the timepoints of the primal.  This
yields an even more memory-efficient implementation where the
derivatives can make use of the same `ts` and `time_matrix` as the
primal.